+++ src/Commands.ts
@@ -392,12 +392,21 @@ export class Commands {
     }
 
     /**
+     * Puts the stash node text from a template to clipboard.
+     *
+     * @param node the involved node
+     */
+    public clipboardFromTemplate = (node: StashNode): void => {
+        void vscode.env.clipboard.writeText(this.stashLabels.clipboardTemplate(node))
+    }
+
+    /**
      * Puts the stash node text on clipboard.
      *
      * @param node the involved node
      */
-    public toClipboard = (node: StashNode): void => {
-        void vscode.env.clipboard.writeText(this.stashLabels.forClipboard(node))
+    public toClipboardFromObject = (node: StashNode): void => {
+        void vscode.env.clipboard.writeText(this.stashLabels.clipboardNode(node))
     }
 
     /**